# miletos-be/miletos-go/internal/infra/http/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Miletos Workflow Runtime API",
    "version": "1.0.0",
    "description": "Internal Go runtime API for synchronous and asynchronous workflow execution and execution observability."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8081",
      "description": "Local runtime"
    }
  ],
...
  "paths": {
...
    "/api/v1/plugins": {
      "get": {
        "tags": [
          "Plugins"
        ],
        "operationId": "listPlugins",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestID"
          },
          {
            "$ref": "#/components/parameters/CorrelationID"
          }
        ],
        "responses": {
          "200": {
            "description": "Deterministic plugin catalog",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PluginListResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/api/v1/executions/sync": {
      "post": {
        "tags": [
          "Executions"
        ],
        "operationId": "executeSync",
        "description": "Executes a workflow locally and waits for a terminal domain result.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CompanyID"
          },
          {
            "$ref": "#/components/parameters/RequestID"
          },
          {
            "$ref": "#/components/parameters/CorrelationID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecutionRequestBody"
              },
              "example": {
                "definition": {
                  "id": "workflow-sync-example",
                  "name": "Sync Example",
                  "revision": 1,
                  "nodes": [
                    {
                      "id": "source",
                      "pluginType": "core.static-input",
                      "pluginVersion": "v1",
                      "configuration": {
                        "value": {
                          "message": "hello"
                        }
                      },
                      "position": {
                        "x": 100,
                        "y": 120
                      }
                    },
                    {
                      "id": "terminal",
                      "pluginType": "core.terminal",
                      "pluginVersion": "v1",
                      "configuration": {},
                      "position": {
                        "x": 420,
                        "y": 120
                      }
                    }
                  ],
                  "edges": [
                    {
                      "id": "edge-1",
                      "sourceNodeId": "source",
                      "sourceOutputPort": "output",
                      "targetNodeId": "terminal",
                      "targetInputPort": "input"
                    }
                  ],
                  "metadata": {
                    "purpose": "swagger-example"
                  }
                },
                "initialVariables": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Terminal synchronous execution result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/api/v1/executions/async": {
      "post": {
        "tags": [
          "Executions"
        ],
        "operationId": "executeAsync",
        "description": "Durably accepts asynchronous workflow execution after persistence and initial outbox creation.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CompanyID"
          },
          {
            "$ref": "#/components/parameters/RequestID"
          },
          {
            "$ref": "#/components/parameters/CorrelationID"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecutionRequestBody"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Asynchronous execution durably accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/api/v1/executions": {
      "get": {
        "tags": [
          "Executions"
        ],
        "operationId": "listExecutions",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CompanyID"
          },
          {
            "$ref": "#/components/parameters/RequestID"
          },
          {
            "$ref": "#/components/parameters/CorrelationID"
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/After"
          },
          {
            "$ref": "#/components/parameters/WorkflowID"
          },
          {
            "$ref": "#/components/parameters/ExecutionStatus"
          }
        ],
        "responses": {
          "200": {
            "description": "Tenant-scoped execution page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionPageResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/api/v1/executions/{executionId}": {
      "get": {
        "tags": [
          "Executions"
        ],
        "operationId": "getExecution",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ExecutionID"
          },
          {
            "$ref": "#/components/parameters/CompanyID"
          },
          {
            "$ref": "#/components/parameters/RequestID"
          },
          {
            "$ref": "#/components/parameters/CorrelationID"
          }
        ],
        "responses": {
          "200": {
            "description": "Execution detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionSummaryResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          }
        }
      }
    },
    "/api/v1/executions/{executionId}/definition": {
      "get": {
        "tags": [
          "Observability"
        ],
        "operationId": "getExecutionDefinition",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ExecutionID"
          },
          {
            "$ref": "#/components/parameters/CompanyID"
          },
          {
            "$ref": "#/components/parameters/RequestID"
          },
          {
            "$ref": "#/components/parameters/CorrelationID"
          }
        ],
        "responses": {
          "200": {
            "description": "Immutable definition snapshot",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionDefinitionResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          }
        }
      }
    },
    "/api/v1/executions/{executionId}/nodes": {
      "get": {
        "tags": [
          "Observability"
        ],
        "operationId": "listExecutionNodes",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ExecutionID"
          },
          {
            "$ref": "#/components/parameters/CompanyID"
          },
          {
            "$ref": "#/components/parameters/RequestID"
          },
          {
            "$ref": "#/components/parameters/CorrelationID"
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/After"
          }
        ],
        "responses": {
          "200": {
            "description": "Node execution page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeExecutionPageResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          }
        }
      }
    },
    "/api/v1/executions/{executionId}/events": {
      "get": {
        "tags": [
          "Observability"
        ],
        "operationId": "listExecutionEvents",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ExecutionID"
          },
          {
            "$ref": "#/components/parameters/CompanyID"
          },
          {
            "$ref": "#/components/parameters/RequestID"
          },
          {
            "$ref": "#/components/parameters/CorrelationID"
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/After"
          }
        ],
        "responses": {
          "200": {
            "description": "Execution event page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionEventPageResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          }
        }
      }
    },
    "/api/v1/executions/{executionId}/logs": {
      "get": {
        "tags": [
          "Observability"
        ],
        "operationId": "listExecutionLogs",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ExecutionID"
          },
          {
            "$ref": "#/components/parameters/CompanyID"
          },
          {
            "$ref": "#/components/parameters/RequestID"
          },
          {
            "$ref": "#/components/parameters/CorrelationID"
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/After"
          }
        ],
        "responses": {
          "200": {
            "description": "Safe execution log page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionLogPageResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          }
        }
      }
    },
    "/api/v1/executions/{executionId}/errors": {
      "get": {
        "tags": [
          "Observability"
        ],
        "operationId": "listExecutionErrors",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ExecutionID"
          },
          {
            "$ref": "#/components/parameters/CompanyID"
          },
          {
            "$ref": "#/components/parameters/RequestID"
          },
          {
            "$ref": "#/components/parameters/CorrelationID"
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/After"
          }
        ],
        "responses": {
          "200": {
            "description": "Structured safe execution error page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionErrorPageResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          }
        }
      }
    },
    "/api/v1/executions/{executionId}/recover": {
      "post": {
        "tags": [
          "Executions"
        ],
        "operationId": "recoverExecution",
        "description": "Creates an idempotent child execution that preserves successful source nodes and schedules only safe failed or timed-out recovery boundaries. The terminal source execution is never reopened.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ExecutionID"
          },
          {
            "$ref": "#/components/parameters/CompanyID"
          },
          {
            "$ref": "#/components/parameters/RequestID"
          },
          {
            "$ref": "#/components/parameters/CorrelationID"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "responses": {
          "202": {
            "description": "Partial recovery accepted or replayed from the same idempotency key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartialRecoveryResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Internal service token supplied by the trusted control-plane caller. The credential is never embedded in this specification."
      }
    },
    "parameters": {
      "CompanyID": {
        "name": "X-Miletos-Company-ID",
        "in": "header",
        "required": true,
        "description": "Trusted tenant context supplied after internal service authentication. It is not end-user authorization.",
        "schema": {
          "type": "string",
          "minLength": 1
        }
      },
      "RequestID": {
        "name": "X-Request-ID",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
      "CorrelationID": {
        "name": "X-Correlation-ID",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "description": "Company-scoped idempotency key for asynchronous execution.",
        "schema": {
          "type": "string",
          "minLength": 1
        }
      },
      "ExecutionID": {
        "name": "executionId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 1
        }
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 50
        }
      },
      "After": {
        "name": "after",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
      "WorkflowID": {
        "name": "workflowId",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
      "ExecutionStatus": {
        "name": "status",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "enum": [
            "CREATED",
            "VALIDATING",
            "REJECTED",
            "QUEUED",
            "RUNNING",
            "SUCCEEDED",
            "FAILED",
            "CANCELLED",
            "TIMED_OUT"
          ]
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/APIErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid internal service authentication",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/APIErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found or not visible in the trusted company scope",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/APIErrorResponse"
            }
          }
        }
      },
      "Conflict": {
        "description": "Idempotency conflict",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/APIErrorResponse"
            }
          }
        }
      },
      "MethodNotAllowed": {
        "description": "HTTP method is not supported",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/APIErrorResponse"
            }
          }
        }
      },
      "PayloadTooLarge": {
        "description": "Request body exceeds the configured limit",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/APIErrorResponse"
            }
          }
        }
      },
      "UnsupportedMediaType": {
        "description": "Content-Type must be application/json",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/APIErrorResponse"
            }
          }
        }
      },
      "ServiceUnavailable": {
        "description": "Required runtime dependency is unavailable",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/APIErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
...
      "ExecutionRequestBody": {
        "type": "object",
        "required": [
          "definition"
        ],
        "properties": {
          "definition": {
            "$ref": "#/components/schemas/WorkflowDefinitionRequest"
          },
          "initialVariables": {
            "type": "object",
            "additionalProperties": {}
          }
        }
      },
      "ExecutionResponse": {
        "type": "object",
        "required": [
          "executionId",
          "workflowId",
          "workflowRevision",
          "mode",
          "status",
          "correlationId",
          "createdAt"
        ],
        "properties": {
          "executionId": {
            "type": "string"
          },
          "workflowId": {
            "type": "string"
          },
          "workflowRevision": {
            "type": "integer",
            "format": "int64"
          },
          "mode": {
            "type": "string",
            "enum": [
              "SYNC",
              "ASYNC"
            ]
          },
          "status": {
            "type": "string"
          },
          "correlationId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "finishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "scheduledRoots": {
            "type": "integer"
          }
        }
      },
...
      "ExecutionSummaryResponse": {
        "type": "object",
        "required": [
          "executionId",
          "workflowId",
          "workflowRevision",
          "mode",
          "status",
          "correlationId",
          "createdAt",
          "updatedAt",
          "isStalled"
        ],
        "properties": {
          "executionId": {
            "type": "string"
          },
          "workflowId": {
            "type": "string"
          },
          "workflowRevision": {
            "type": "integer",
            "format": "int64"
          },
          "mode": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "correlationId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "validatingAt": {
            "type": "string",
            "format": "date-time"
          },
          "queuedAt": {
            "type": "string",
            "format": "date-time"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "finishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "isStalled": {
            "type": "boolean"
          }
        }
      },
      "ExecutionPageResponse": {
        "type": "object",
        "required": [
          "items",
          "next",
          "hasNext"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExecutionSummaryResponse"
            }
          },
          "next": {
            "type": "string"
          },
          "hasNext": {
            "type": "boolean"
          }
        }
      },
...
      "NodeExecutionPageResponse": {
        "type": "object",
        "required": [
          "items",
          "next",
          "hasNext"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NodeExecutionResponse"
            }
          },
          "next": {
            "type": "string"
          },
          "hasNext": {
            "type": "boolean"
          }
        }
      },
...
      "PluginListResponse": {
        "type": "object",
        "required": [
          "items",
          "count"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PluginDescriptor"
            }
          },
          "count": {
            "type": "integer"
          }
        }
      }
    }
  }
}